Fix for "find" when iterating folders without permission
When "find" is used to look for the mainline debian package, it looks in
/home and /root. If there is any folder where find has no permission to
look into (e.g. 'thinclient_drives' when using xrdp), the command will
fail.

This is because 'find' will return with code 1 when there is any error
(related or not).
This patch will first discard all errors written on stderr (send it to
/dev/null) and second will always return 0 exit code to prevent the
script from failing.

The exit code is not relevant in this case because the result of found
mainline-kernel packages is checked (has to be exactly 1 result).
If there are more (or no) matches, script will still fail.